pyproject.toml
[project]
name = "commee"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Dmitry Aminev", email = "voltterra@users.noreply.github.com" }
]
requires-python = ">=3.13"
dependencies = [
    "click>=8.3.1",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
]

[project.scripts]
commee = "commee.test_cli:cli"

[build-system]
requires = ["uv_build>=0.9.14,<0.10.0"]
build-backend = "uv_build"
